+++ app/src/main/java/com/example/foodii/feature/apifoodii/meal/domain/entity/DailySummary.kt
@@ -2,6 +2,6 @@ package com.example.foodii.feature.apifoodii.meal.domain.entity
 
 data class DailySummary(
     val date: String,
-    val totalCalories: Int,
+    val totalCalories: Double,
     val meals: List<FoodiiMeal>
 )
